display eprints, dois, urls, as secondary bulleted list items
