+++ src/literature/components/ReferenceItem/index.js
@@ -1,0 +1,3 @@
+import ReferenceItem from './ReferenceItem';
+
+export default ReferenceItem;
