+++ src/literature/components/__tests__/ArxivEprintList.test.jsx
@@ -11,9 +11,19 @@ describe('ArxivEprintList', () => {
         value: '123.12345',
       },
     ]);
-    const wrapper = shallow((
-      <ArxivEprintList eprints={eprints} />
-    ));
+    const wrapper = shallow(<ArxivEprintList eprints={eprints} />);
     expect(wrapper.dive()).toMatchSnapshot();
+  });
+
+  it('renders with wrapperClassName', () => {
+    const eprints = fromJS([
+      {
+        value: '123.12345',
+      },
+    ]);
+    const wrapper = shallow(
+      <ArxivEprintList eprints={eprints} wrapperClassName="test" />
+    );
+    expect(wrapper).toMatchSnapshot();
   });
 });
