+++ src/literature/components/__tests__/DOIList.test.jsx
@@ -6,7 +6,6 @@ import DOIList from '../DOIList';
 
 describe('DOIList', () => {
   it('renders with dois', () => {
-    // TODO assert if value is there! since it is inside function not in snapshot.
     const dois = fromJS([
       { value: '12.1234/1234567890123_1234' },
       {
@@ -16,5 +15,11 @@ describe('DOIList', () => {
     ]);
     const wrapper = shallow(<DOIList dois={dois} />);
     expect(wrapper.dive()).toMatchSnapshot();
+  });
+
+  it('renders with wrapperClassName', () => {
+    const dois = fromJS([{ value: '12.1234/1234567890123_1234' }]);
+    const wrapper = shallow(<DOIList dois={dois} wrapperClassName="test" />);
+    expect(wrapper).toMatchSnapshot();
   });
 });
